Add a development mode which allows CORS.
