+++ src/main/java/science/icebreaker/config/WebConfig.java
@@ -1,0 +1,49 @@
+package science.icebreaker.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+
+@Configuration
+@EnableSwagger2
+public class WebConfig {
+
+    private final boolean development;
+
+
+    public WebConfig(
+            @Value("${icebreaker.development}") boolean development
+    ) {
+        this.development = development;
+    }
+
+
+    public Docket restAPI() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.ant("/*"))
+                .build();
+    }
+
+
+    @Bean
+    public WebMvcConfigurer corsConfigurer() {
+        return new WebMvcConfigurer() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                if (development) {
+                    registry.addMapping("/**").allowedOrigins("*");
+                }
+            }
+        };
+    }
+}
