+++ src/main/resources/application.properties
@@ -8,5 +8,6 @@ spring.jpa.show-sql=true
 spring.jpa.generate-ddl=false
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+icebreaker.development=false
 icebreaker.jwt-secret=secret
 icebreaker.jwt-token-validity-ms=259200000
